Add default routing to macOS
