+++ client/internal/routemanager/routemanager.go
@@ -13,9 +13,15 @@ import (
 	nbnet "github.com/netbirdio/netbird/util/net"
 )
 
+type ref struct {
+	count   int
+	nexthop netip.Addr
+	intf    string
+}
+
 type RouteManager struct {
-	// refCountMap keeps track of the reference count for prefixes
-	refCountMap map[netip.Prefix]int
+	// refCountMap keeps track of the reference ref for prefixes
+	refCountMap map[netip.Prefix]ref
 	// prefixMap keeps track of the prefixes associated with a connection ID for removal
 	prefixMap   map[nbnet.ConnectionID][]netip.Prefix
 	addRoute    AddRouteFunc
@@ -23,13 +29,13 @@ type RouteManager struct {
 	mutex       sync.Mutex
 }
 
-type AddRouteFunc func(prefix netip.Prefix) error
-type RemoveRouteFunc func(prefix netip.Prefix) error
+type AddRouteFunc func(prefix netip.Prefix) (nexthop netip.Addr, intf string, err error)
+type RemoveRouteFunc func(prefix netip.Prefix, nexthop netip.Addr, intf string) error
 
 func NewRouteManager(addRoute AddRouteFunc, removeRoute RemoveRouteFunc) *RouteManager {
 	// TODO: read initial routing table into refCountMap
 	return &RouteManager{
-		refCountMap: map[netip.Prefix]int{},
+		refCountMap: map[netip.Prefix]ref{},
 		prefixMap:   map[nbnet.ConnectionID][]netip.Prefix{},
 		addRoute:    addRoute,
 		removeRoute: removeRoute,
@@ -40,17 +46,22 @@ func (rm *RouteManager) AddRouteRef(connID nbnet.ConnectionID, prefix netip.Pref
 	rm.mutex.Lock()
 	defer rm.mutex.Unlock()
 
-	log.Debugf("Increasing route ref count %d for prefix %s", rm.refCountMap[prefix], prefix)
+	ref := rm.refCountMap[prefix]
+	log.Debugf("Increasing route ref count %d for prefix %s", ref.count, prefix)
 
 	// Add route to the system, only if it's a new prefix
-	if rm.refCountMap[prefix] == 0 {
+	if ref.count == 0 {
 		log.Debugf("Adding route for prefix %s", prefix)
-		if err := rm.addRoute(prefix); err != nil {
+		nexthop, intf, err := rm.addRoute(prefix)
+		if err != nil {
 			return fmt.Errorf("failed to add route for prefix %s: %w", prefix, err)
 		}
+		ref.nexthop = nexthop
+		ref.intf = intf
 	}
 
-	rm.refCountMap[prefix]++
+	ref.count++
+	rm.refCountMap[prefix] = ref
 	rm.prefixMap[connID] = append(rm.prefixMap[connID], prefix)
 
 	return nil
@@ -68,17 +79,19 @@ func (rm *RouteManager) RemoveRouteRef(connID nbnet.ConnectionID) error {
 
 	var result *multierror.Error
 	for _, prefix := range prefixes {
-		log.Debugf("Decreasing route ref count %d for prefix %s", rm.refCountMap[prefix], prefix)
-		if rm.refCountMap[prefix] == 1 {
+		ref := rm.refCountMap[prefix]
+		log.Debugf("Decreasing route ref count %d for prefix %s", ref.count, prefix)
+		if ref.count == 1 {
 			log.Debugf("Removing route for prefix %s", prefix)
 			// TODO: don't fail if the route is not found
-			if err := rm.removeRoute(prefix); err != nil {
+			if err := rm.removeRoute(prefix, ref.nexthop, ref.intf); err != nil {
 				result = multierror.Append(result, fmt.Errorf("remove route for prefix %s: %w", prefix, err))
 				continue
 			}
 			delete(rm.refCountMap, prefix)
 		} else {
-			rm.refCountMap[prefix]--
+			ref.count--
+			rm.refCountMap[prefix] = ref
 		}
 	}
 	delete(rm.prefixMap, connID)
@@ -94,11 +107,12 @@ func (rm *RouteManager) Flush() error {
 	var result *multierror.Error
 	for prefix := range rm.refCountMap {
 		log.Debugf("Removing route for prefix %s", prefix)
-		if err := rm.removeRoute(prefix); err != nil {
+		ref := rm.refCountMap[prefix]
+		if err := rm.removeRoute(prefix, ref.nexthop, ref.intf); err != nil {
 			result = multierror.Append(result, fmt.Errorf("remove route for prefix %s: %w", prefix, err))
 		}
 	}
-	rm.refCountMap = map[netip.Prefix]int{}
+	rm.refCountMap = map[netip.Prefix]ref{}
 	rm.prefixMap = map[nbnet.ConnectionID][]netip.Prefix{}
 
 	return result.ErrorOrNil()
