+++ client/internal/routemanager/systemops_android.go
@@ -3,6 +3,9 @@ package routemanager
 import (
 	"net"
 	"net/netip"
+	"runtime"
+
+	log "github.com/sirupsen/logrus"
 
 	"github.com/netbirdio/netbird/client/internal/peer"
 	"github.com/netbirdio/netbird/iface"
@@ -16,10 +19,15 @@ func cleanupRouting() error {
 	return nil
 }
 
-func addToRouteTableIfNoExists(prefix netip.Prefix, addr, intf string) error {
+func enableIPForwarding() error {
+	log.Infof("Enable IP forwarding is not implemented on %s", runtime.GOOS)
 	return nil
 }
 
-func removeFromRouteTableIfNonSystem(prefix netip.Prefix, addr, intf string) error {
+func addVPNRoute(netip.Prefix, string) error {
+	return nil
+}
+
+func removeVPNRoute(netip.Prefix, string) error {
 	return nil
 }
