--- client/internal/routemanager/systemops_bsd_test.go
@@ -1,10 +1,0 @@
-//go:build darwin || dragonfly || freebsd || netbsd || openbsd
-
-package routemanager
-
-import "testing"
-
-//nolint:unused
-func setupDummyInterfacesAndRoutes(t *testing.T) {
-	t.Helper()
-}
