+++ util/net/dialer_mobile.go
@@ -1,0 +1,6 @@
+//go:build android || ios
+
+package net
+
+func (d *Dialer) init() {
+}
