+++ util/net/listener_mobile.go
@@ -1,0 +1,6 @@
+//go:build android || ios
+
+package net
+
+func (l *ListenerConfig) init() {
+}
